Add genotype metadata
This commit adds genotype metadata that is reported on NCBI to the ingest output, and allows visualization of that information on the genome tree and the N450 tree.

Many samples have genotype information reported within the `virus-name` field of the NDJSON record from NCBI. This commit uses a custom script to parse the genotypes from that field and outputs the parsed genotypes to a column named `genotype_ncbi`.

Genotypes are not available on NCBI for the vaccine strains or WHO genotype reference strains, and so this commit manually adds the genotypes for those strains. Information for WHO genotype reference strains is from https://www.who.int/publications/i/item/WER8709