Ensure all text files (except ingest/vendored) have end-of-file newlines
A long-standing Unix convention that when broken can lead to confusing
and incorrect behaviour from many commands when they read such files.

The offending files in ingest/vendored will be fixed when we next update
it from upstream.  Updating it now would necessitate a slew of unrelated
changes to this repo.