Add `nextstrain run` instructions to READMEs
I skipped the nextclade/README.md because the instructions there are more 
"developer-y" docs for creating the Nextclade reference tree rather than 
how to run the workflow. 

In hindsight, I'm not entirely sure the nextclade workflow needs to be supported 
by `nextstrain run` since it's usually such an iterative process that requires
updating the source.