Consolidate filepath resolution
Filepaths in the config are either resolved within rules or at the start
of the workflow.

Resolve all filepaths at the start of the workflow to keep things
consolidated.

The current implementation does not allow for string placeholders in
filepaths, so the config has been adjusted to replace those with
expanded versions.