Use shared functions
These functions are available in shared/vendored/snakemake/config.smk
which is included earlier in the Snakefile.

I removed the main() function wrapper since there primary purpose of
that was to make the file more readable with high-level function calls
at the top. That's no longer necessary now that the file is comprised
solely of high-level function calls.